(rgbd) fix jazzy/rolling build breakage and misplaced clang_tidy suppressions
The previous clang_tidy fix commit added literal #include <image_geometry/pinhole_camera_model.h>
and <message_filters/subscriber.h>/<synchronizer.h> to several .cpp files. Those headers only exist
as .h on humble; jazzy/rolling-u24 ship .hpp only (or, for message_filters on rolling, no .h shim at
all), which broke the build on those distros ("No such file or directory"). Wrap them in the same
__has_include(...hpp) guard already used elsewhere in this codebase (client_ros.h, view.cpp), with
"IWYU pragma: keep" to stop misc-include-cleaner from flagging the guarded include as unused.

Also fix four NOLINT suppressions added by that same commit that never actually suppressed anything:
NOLINTNEXTLINE only covers the single following line, but each was followed by a second comment line
before the flagged statement, or (for the UninitializedObject finding in rgbd_to_shm.cpp) placed
above the constructor declaration while clang-analyzer anchors that diagnostic to the constructor's
last statement. Moved each NOLINTNEXTLINE to sit directly above the actual flagged line.

Additionally fixes three findings that had no suppression yet: missing <cstdint> for uint16_t in
serialization.cpp, a double->float narrowing conversion in rgbd_to_pcd.cpp, and a modernize-use-auto
in save.cpp.

Remaining clang_tidy findings (message_filters/subscriber.h virtual-call-in-destructor inside a ROS
system header, and a boost::numeric_cast/PCL/clang-tidy-21 template instantiation crash affecting
server_ros.cpp and rgbd_to_pcd.cpp) are third-party/toolchain issues outside this package.

+++ rgbd/tools/rgbd_to_pcd.cpp
@@ -71,7 +71,7 @@ int main(int argc, char** argv)
                 // union; there is no non-union accessor.
                 cloud.at(jj, ii).x = static_cast<float>((-half_width + j) * d / fx);
                 cloud.at(jj, ii).y = static_cast<float>((-half_height + i) * d / fy);
-                cloud.at(jj, ii).z = d;
+                cloud.at(jj, ii).z = static_cast<float>(d);
                 cloud.at(jj, ii).r = bgr[2];
                 cloud.at(jj, ii).g = bgr[1];
                 cloud.at(jj, ii).b = bgr[0];
